Move usePerNetworkCustomization to hooks directory
