+++ projects/packages/publicize/_inc/components/customize-and-preview/tab-panels/desktop/index.tsx
@@ -1,6 +1,6 @@
 import { TabPanel } from '@wordpress/components';
 import { useCallback } from 'react';
-import { usePerNetworkCustomization } from '../../customization-toggle/use-per-network-customization';
+import { usePerNetworkCustomization } from '../../../../hooks/use-per-network-customization';
 import { ConnectionTab } from '../types';
 import { useConnectionTabs } from '../use-connection-tabs';
 import styles from './styles.module.scss';
